requirements.txt
flask==2.2.2
gunicorn==20.1.0
spacy==3.5.0
firebase_admin==6.0.1
google-cloud-firestore==2.9.1
python-dotenv==0.21.1
detoxify==0.5.1
numpy==1.24.1